sub acts act and stream lined sequential list and required all / any
