+++ sharpy/plans/__init__.py
@@ -2,3 +2,4 @@ from .build_order import BuildOrder
 from .build_order import Step
 from .step_gas import StepBuildGas
 from .sequential_list import SequentialList
+from .sub_acts import SubActs
